rewrite project README
implement fee-on minting (1/6th of Δsqrt(k))
update kLast on mint/burn when fees are on

+++ src/interfaces/IV2Callee.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.30;
 
 interface IV2Callee {
-    function V2Call(address sender, uint amount0, uint amount1, bytes calldata data) external;
+    function V2Call(address sender, uint256 amount0, uint256 amount1, bytes calldata data) external;
 }
